+++ src/pages/programsWip/form/EnrollmentFormFormContents.tsx
@@ -1,27 +1,24 @@
 import i18n from '@dhis2/d2-i18n'
-import { IconInfo16, NoticeBox, Tab, TabBar } from '@dhis2/ui'
 import React, { useState } from 'react'
 import { useField, UseFieldConfig } from 'react-final-form'
-import { useParams } from 'react-router-dom'
 import {
     SectionedFormSection,
     StandardFormSectionDescription,
     StandardFormSectionTitle,
 } from '../../../components'
-import { TooltipWrapper } from '../../../components/tooltip'
-import { DataSet } from '../../../types/generated'
-import { ProgramsFromFilters } from '../Edit'
-import classes from './EnrollmentFormFormContent.module.css'
-import formType = DataSet.formType
+import { SectionFormSectionsList } from '../../../components/formCreators/SectionFormList'
+import {
+    FormType,
+    TabbedFormTypePicker,
+} from '../../../components/formCreators/TabbedFormTypePicker'
+import { SchemaName } from '../../../types'
+import { ProgramValues } from '../Edit'
+import { EditOrNewEnrollmentSectionForm } from './sectionForm/EntrollmentSectionForm'
 
-const disabledFormTypeText = i18n.t(
-    'Program must be saved before this form type can be configured.'
-)
-
-const useProgramField = <T extends keyof ProgramsFromFilters>(
+const useProgramField = <T extends keyof ProgramValues>(
     name: T,
-    options?: UseFieldConfig<ProgramsFromFilters[T], ProgramsFromFilters[T]>
-) => useField<ProgramsFromFilters[T]>(name, options)
+    options?: UseFieldConfig<ProgramValues[T], ProgramValues[T]>
+) => useField<ProgramValues[T]>(name, options)
 
 export const EnrollmentFormFormContents = React.memo(function FormFormContents({
     name,
@@ -33,18 +30,9 @@ export const EnrollmentFormFormContents = React.memo(function FormFormContents({
     const trackedEntityAttributes = useProgramField(
         'programTrackedEntityAttributes'
     ).input.value
-    const modelId = useParams().id
-    const isCreatingNewProgram = !modelId
-    const [selectedFormType, setSelectedFormType] = useState<formType>(
-        formType.DEFAULT
+    const [selectedFormType, setSelectedFormType] = useState<FormType>(
+        FormType.DEFAULT
     )
-    const androidFormType =
-        sections && sections.length > 0
-            ? i18n.t('Section form')
-            : i18n.t('Basic form')
-
-    const webFormType = dataEntryForm ? i18n.t('Custom form') : androidFormType
-
     return (
         <SectionedFormSection name={name}>
             <StandardFormSectionTitle>
@@ -55,117 +43,23 @@ export const EnrollmentFormFormContents = React.memo(function FormFormContents({
                     'Choose and configure the type of form that will be used to collect information during enrollment.'
                 )}
             </StandardFormSectionDescription>
-
-            {!isCreatingNewProgram && trackedEntityAttributes.length > 0 && (
-                <NoticeBox
-                    title={i18n.t('Form type based on your current setup')}
-                    className={classes.formTypeInfo}
-                >
-                    {i18n.t(
-                        'Web will display {{webFormType}}  |  Android will display {{androidFormType}}',
-                        {
-                            webFormType,
-                            androidFormType,
-                        }
-                    )}
-                </NoticeBox>
-            )}
-            <div
-                className={classes.formTypeTabsContainer}
-                onClick={(e) => {
-                    e.preventDefault()
-                }}
+            <TabbedFormTypePicker
+                sectionsLength={sections.length}
+                hasDataEntryForm={!!dataEntryForm}
+                hasDataToDisplay={trackedEntityAttributes.length > 0}
+                onFormTypeChange={setSelectedFormType}
+                selectedFormType={selectedFormType}
             >
-                <TabBar fixed>
-                    <Tab
-                        onClick={(_, event) => {
-                            event.preventDefault()
-                            setSelectedFormType(formType.DEFAULT)
-                        }}
-                        selected={selectedFormType === formType.DEFAULT}
-                    >
-                        <div className={classes.formTypeTab}>
-                            {i18n.t('Basic form')}
-                            <TooltipWrapper
-                                condition
-                                className={classes.infoTooltipWrapper}
-                                content={i18n.t(
-                                    'Basic forms display an auto-generated list of attributes. \n' +
-                                        'They will be displayed only if no Section or Custom form is created.'
-                                )}
-                            >
-                                <IconInfo16 />
-                            </TooltipWrapper>
-                        </div>
-                    </Tab>
-                    <Tab
-                        onClick={(_, event) => {
-                            event.preventDefault()
-                            setSelectedFormType(formType.SECTION)
-                        }}
-                        disabled={isCreatingNewProgram}
-                        selected={selectedFormType === formType.SECTION}
-                    >
-                        <TooltipWrapper
-                            condition={isCreatingNewProgram}
-                            content={disabledFormTypeText}
-                        >
-                            <div className={classes.formTypeTab}>
-                                {i18n.t('Section form')}
-                                <TooltipWrapper
-                                    condition={!isCreatingNewProgram}
-                                    className={classes.infoTooltipWrapper}
-                                    content={i18n.t(
-                                        'Section forms let you create sections to organize data items for easier entry. If a Section form is created, it will be shown instead of the Basic form.'
-                                    )}
-                                >
-                                    <IconInfo16 />
-                                </TooltipWrapper>
-                            </div>
-                        </TooltipWrapper>
-                    </Tab>
-
-                    <Tab
-                        onClick={(_, event) => {
-                            event.preventDefault()
-                            setSelectedFormType(formType.CUSTOM)
-                        }}
-                        disabled={isCreatingNewProgram}
-                        selected={selectedFormType === formType.CUSTOM}
-                    >
-                        <TooltipWrapper
-                            condition={isCreatingNewProgram}
-                            content={disabledFormTypeText}
-                        >
-                            <div
-                                className={classes.formTypeTab}
-                                onClick={(e) => {
-                                    e.preventDefault()
-                                }}
-                            >
-                                {i18n.t('Custom form')}
-                                <TooltipWrapper
-                                    condition={!isCreatingNewProgram}
-                                    className={classes.infoTooltipWrapper}
-                                    content={i18n.t(
-                                        'Custom forms let you design a fully customized layout for data entry. \n' +
-                                            '\n' +
-                                            'On Web, Custom forms (if created) will always be shown.\n' +
-                                            'Android does not support Custom forms; \n' +
-                                            'instead, the Section form will be shown (if created), else the Basic form.'
-                                    )}
-                                >
-                                    <IconInfo16 />
-                                </TooltipWrapper>
-                            </div>
-                        </TooltipWrapper>
-                    </Tab>
-                </TabBar>
-            </div>
-            <div className={classes.formTypeTabsContent}>
-                {selectedFormType === formType.SECTION && <>TODO</>}
-                {selectedFormType === formType.CUSTOM && <>TODO</>}
-            </div>
+                {selectedFormType === FormType.SECTION && (
+                    <SectionFormSectionsList
+                        sectionsFieldName={'programSections'}
+                        SectionFormComponent={EditOrNewEnrollmentSectionForm}
+                        schemaName={SchemaName.programSection}
+                        otherProps={{ sectionsLength: sections.length }}
+                    />
+                )}
+                {selectedFormType === FormType.CUSTOM && <>TODO</>}
+            </TabbedFormTypePicker>
         </SectionedFormSection>
     )
 })